\name{atlas.rois}
\alias{atlas.rois}
\title{ Generate Table Combining Cluster Map and Neuroanatomical Atlas}
\description{}
\usage{
  atlas.rois(cluster.nii, cluster.vol="all", atlas.nii, atlas.csv, save.dir, prefix)
}
\arguments{
  \item{cluster.nii}{}
  \item{cluster.vol}{}
  \item{atlas.nii}{}
  \item{atlas.csv}{}
  \item{save.dir}{}
  \item{prefix}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
