\name{cluster.1.vol.n}
\alias{cluster.1.vol.n}
\title{Convert Single Volume to Multivolume}
\description{}
\usage{
  cluster.1.vol.n(nii.file, save.dir=NULL, file.name=NULL)
}
\arguments{
  \item{nii.file}{}
  \item{save.dir}{}
  \item{file.name}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+cluster@gmail.com>
}
\examples{}