\name{cluster.p}
\alias{cluster.p}
\title{
Cluster P Maps
}
\description{}
\usage{
cluster.p(nii.p, vol.p=1, nii.sign=NULL, vol.sign=vol.p, nii.mask=NULL, vol.mask=vol.p, p.thresh=0.001, cluster.size, save.dir, connectivity=NULL)
}
\arguments{
  \item{nii.p}{}
  \item{vol.p}{}
  \item{nii.sign}{}
  \item{vol.sign}{}
  \item{nii.mask}{}
  \item{vol.mask}{}
  \item{p.thresh}{}
  \item{cluster.size}{}
  \item{save.dir}{}
  \item{connectivity}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+cluster@gmail.com>
}
\examples{}