\name{split.cluster.anatomy}
\alias{split.cluster.anatomy}
\title{Split Cluster mask by Anatomical Regions}
\description{}
\usage{
split.cluster.anatomy(data.cluster, data.anatomy, which.cluster="all", cluster.sz)
}
\arguments{
  \item{data.cluster}{}
  \item{data.anatomy}{}
  \item{which.cluster}{}
  \item{cluster.sz}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}